feat: add macOS support and harden cross-platform memory scanner
Major release adding native macOS support, fixing latent bugs in the
Windows/Linux backends, and tightening cross-platform robustness. See
CHANGELOG.md for the full list.

Added
- macOS backend via Mach VM APIs (task_for_pid, mach_vm_read_overwrite,
  mach_vm_write, mach_vm_region, mach_vm_protect). Self-process works
  without entitlements; cross-process needs com.apple.security.cs.debugger
  or SIP off + root.
- snapshot_memory_regions() + memory_regions= kwarg on search_by_value*
  and search_by_addresses for refine-scan workflows.
- bufflength is now optional for numeric types (int->4, float->8, bool->1).
- iter_region_chunks reads multi-GB regions in 256MB chunks across all
  three backends (prevents OOM in browser/JVM-sized targets).
- PyMemoryEditorError base + AmbiguousProcessNameError.
- py.typed marker; mypy + pytest-cov in CI.

Fixed (critical)
- Platform detection: "win" in sys.platform matched darwin, breaking
  macOS imports.
- Read/Write/process_vm_*v now check argtypes/return value; failed reads
  no longer return zeroed buffers indistinguishable from real data.
- scan_memory off-by-one (skipped last value of each region).
- scan_memory_for_exact_value NOT_EXACT_VALUE no longer yields every
  non-matching byte; aligned to target_value_size.
- WindowsProcess default permission: PROCESS_VM_READ instead of
  PROCESS_ALL_ACCESS (least privilege).
- Permission gate now requires VM_READ explicit OR all PROCESS_ALL_ACCESS
  bits set (was passing on any single bit).
- ProcessOperationsEnum.PROCESS_TERMINATE was 0x0800 — same as
  PROCESS_SUSPEND_RESUME — making it a silent Enum alias. Corrected to
  0x0001 per MSDN.
- Linux MEMORY_BASIC_INFORMATION fields widened to 64-bit (regions > 4GB
  no longer truncate).
- Linux /proc/<pid>/maps inode parsed as decimal (was hex).
- Windows MEMORY_BASIC_INFORMATION layout picked per target via
  IsWow64Process (no more corruption when 64-bit Python attaches to
  32-bit target).
- macOS write_process_memory to a read-only page now transparently
  elevates protection via mach_vm_protect and restores it.
- Linux scan filters shared mappings (parity with Win32/macOS).
- read_process_memory(addr, str, n) now decodes with errors="replace"
  to match convert_from_byte_array.

Performance
- 6-8x speedup on numeric scans via struct.iter_unpack + inlined
  comparison loops per scan_type. NOT_EXACT_VALUE overlap check is now
  O(log m) via bisect_left.

Tooling
- CI: 3 OSes (ubuntu/windows/macos) x 6 Pythons (3.8-3.13), flake8 gate,
  mypy informational, pytest-cov.
- Conventional Commits enforced on PR title (lint-pr-title.yml).
- Dependabot config (open-pull-requests-limit: 0; security alerts still
  fire).
- Auto-delete head branch on PR close.
- Tk sample now requires Tk >= 8.6 and aborts with platform-specific
  install hints when missing or outdated.

Removed
- Unused PyMemoryEditor.linux.ptrace package.
- Unused PyMemoryEditor.util.search (KMP/BMH never used in scan path).
- Python 3.6 and 3.7 support (minimum is now 3.8).

+++ PyMemoryEditor/linux/types.py
@@ -6,18 +6,20 @@
 # Read more about iovec here:
 # https://man7.org/linux/man-pages/man3/iovec.3type.html
 
-from ctypes import Structure, c_char_p, c_size_t, c_uint, c_void_p
+from ctypes import Structure, c_char_p, c_size_t, c_uint, c_uint64, c_void_p
 
 
 class MEMORY_BASIC_INFORMATION(Structure):
+    # Address/size/offset fields are 64-bit so that mappings beyond 4 GB
+    # (huge pages, large file mmaps) are not silently truncated on x86_64.
     _fields_ = [
-        ("BaseAddress", c_uint),
-        ("RegionSize", c_uint),
+        ("BaseAddress", c_uint64),
+        ("RegionSize", c_uint64),
         ("Privileges", c_char_p),
-        ("Offset", c_uint),
+        ("Offset", c_uint64),
         ("MajorID", c_uint),
         ("MinorID", c_uint),
-        ("InodeID", c_uint),
+        ("InodeID", c_uint64),
         ("Path", c_char_p),
     ]
